requirements.txt
python>=3.5.0
pyramid>=1.6.1
feedgen>=0.3.2
PyYAML>=3.11
requests>=2.9.1
